Organize Go ABI package by responsibility
Separate the ABI model, requirement schema, value resolution, DWARF
extraction, and structural validation into focused files. Keeping these
concerns distinct makes the versioned schema easier to change without
turning the package entry point into a collection of unrelated helpers.

Mirror those boundaries in the tests and keep the generated type-size
marker private because it is an implementation detail with no package
consumers.
